Optional VEP. VEP S3 documentation.  Jupyter startup correction.

+++ packer/scripts/vep_install.sh
@@ -18,13 +18,13 @@ GSUTIL_PROFILE="/etc/profile.d/gsutil.sh"
 GSUTIL_SOURCE="https://storage.googleapis.com/pub/gsutil.tar.gz"
 GSUTIL_TARGET_DIR="/opt"
 REPOSITORY_URL="https://github.com/Ensembl/ensembl-vep.git"
-S3_MAX_CONCURRENT_REQUESTS="30"
 export VEP_CACHE_DIR="/opt/vep/cache"
 export VEP_S3_SOURCE="s3://$HAIL_BUCKET"
 export VEP_S3_CACHE_PATH="/vep/cache"
 export VEP_S3_LOFTEE_PATH="/vep/loftee_data"
 export VEP_SPECIES="homo_sapiens"
 export VEP_DIR="/opt/vep"
+export PATH="$PATH:/usr/local/bin"
 
 function install_prereqs {
     yum -y install \
@@ -57,16 +57,13 @@ function vep_install {
     mkdir -p "$VEP_CACHE_DIR"
 
     # Confirm cache exists in S3
-    echo -e "Confirming that the VEP cache exists in $HAIL_BUCKET
-        $VEP_S3_SOURCE$VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh37/info.txt
-        $VEP_S3_SOURCE$VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh38/info.txt"
+    echo -e "Confirming that the VEP cache exists in $HAIL_BUCKET"
     POP_VEP_S3_CACHE_PATH=$(echo ${VEP_S3_CACHE_PATH#/})  # Remove leading /
-    aws s3api head-object --bucket "$HAIL_BUCKET" --key "$POP_VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh37/info.txt"
-    aws s3api head-object --bucket "$HAIL_BUCKET" --key "$POP_VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh38/info.txt"
-
-    aws configure set default.s3.max_concurrent_requests "$S3_MAX_CONCURRENT_REQUESTS"
-    aws s3 sync "$VEP_S3_SOURCE$VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh37" "$VEP_CACHE_DIR/$VEP_SPECIES/${VEP_VERSION}_GRCh37"
-    aws s3 sync "$VEP_S3_SOURCE$VEP_S3_CACHE_PATH/$VEP_SPECIES/${VEP_VERSION}_GRCh38" "$VEP_CACHE_DIR/$VEP_SPECIES/${VEP_VERSION}_GRCh38"
+    aws s3api head-object --bucket "$HAIL_BUCKET" --key "$POP_VEP_S3_CACHE_PATH/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh37.tar.gz"
+    aws s3api head-object --bucket "$HAIL_BUCKET" --key "$POP_VEP_S3_CACHE_PATH/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh38.tar.gz"
+    # Copy cache archives in, extract, and remove
+    aws s3 cp "$VEP_S3_SOURCE$VEP_S3_CACHE_PATH/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh37.tar.gz" /tmp
+    aws s3 cp "$VEP_S3_SOURCE$VEP_S3_CACHE_PATH/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh38.tar.gz" /tmp
 
     # Install VEP - the earliest version available from GitHub is 87
     if [ "$VEP_VERSION" -ge 87 ]; then
@@ -74,10 +71,18 @@ function vep_install {
         git clone "$REPOSITORY_URL"
         cd ensembl-vep
         git checkout "release/$VEP_VERSION"
-        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache \
+
+        # Auto install (a)pi, (c)ache, and (f)asta
+        tar --directory "$VEP_CACHE_DIR"  -vxf "/tmp/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh37.tar.gz"
+        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
              --AUTO acf --SPECIES "$VEP_SPECIES" --ASSEMBLY GRCh37 --NO_HTSLIB --NO_UPDATE
-        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache \
+        rm "/tmp/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh37.tar.gz"
+
+        # Auto install (c)ache
+        tar --directory "$VEP_CACHE_DIR"  -vxf "/tmp/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh38.tar.gz"
+        perl INSTALL.pl --DESTDIR "$VEP_DIR" --CACHEDIR "$VEP_DIR"/cache --CACHEURL "$VEP_CACHE_DIR" \
              --AUTO c --SPECIES "$VEP_SPECIES" --ASSEMBLY GRCh38 --NO_HTSLIB --NO_UPDATE
+        rm "/tmp/${VEP_SPECIES}_vep_${VEP_VERSION}_GRCh38.tar.gz"
 
         # Plugins are installed to $HOME.  Install all plugins, then move to common location
         perl INSTALL.pl --AUTO p --PLUGINS all --NO_UPDATE
@@ -92,24 +97,21 @@ function vep_install {
         $GSUTIL_TARGET_DIR/gsutil/gsutil -m cp -r gs://hail-common/vep/vep/Plugins "$VEP_DIR"
     fi
 
-    # Loftee for VEP - set paths appropriately in your Hail vep-configuration.json
+    # Loftee for VEP GRCh37 only
     mkdir -p "$VEP_DIR"/loftee_data
-    # CRCh37 in link - does this matter?  Ask Adam Tebbe
-    # loftee_data below is cached in local s3 bucket for speed
-    #aws s3 cp s3://bcbio_nextgen/human_ancestor.fa.gz "$VEP_DIR"/loftee_data
-    #curl https://personal.broadinstitute.org/konradk/loftee_data/GRCh37/phylocsf_gerp.sql.gz --output "$VEP_DIR"/loftee_data/phylocsf_gerp.sql.gz
-    #gunzip "$VEP_DIR"/loftee_data/phylocsf_gerp.sql.gz
-    #curl https://personal.broadinstitute.org/konradk/loftee_data/GRCh37/GERP_scores.final.sorted.txt.gz --output "$VEP_DIR"/loftee_data/GERP_scores.final.sorted.txt.gz
-
     aws s3 sync "$VEP_S3_SOURCE$VEP_S3_LOFTEE_PATH" "$VEP_DIR"/loftee_data
     gunzip "$VEP_DIR"/loftee_data/phylocsf_gerp.sql.gz
 }
 
-install_prereqs
-gsutil_install
-vep_install
+if [ "$VEP_VERSION" != "none" ]; then
+    install_prereqs
+    gsutil_install
+    vep_install
 
-# Cleanup
-rm -rf /root/.cpanm
-rm -rf /root/.vep
-rm -rf /root/ensembl-vep
+    # Cleanup
+    rm -rf /root/.cpanm
+    rm -rf /root/.vep
+    rm -rf /root/ensembl-vep
+else
+    echo "VEP_VERSION environment variable was \"none\".  Skipping VEP installation."
+fi
